app/requirements.txt
Django
celery
django-celery-beat
bs4
djangorestframework
djoser
psycopg2
redis
gunicorn
load_dotenv
